% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freestatemoney-package.R
\docType{package}
\name{freestatemoney-package}
\alias{freestatemoney}
\alias{freestatemoney-package}
\title{freestatemoney: Load and Parse Maryland Campaign Finance Data}
\description{
Provides functions to download, parse, and analyze campaign finance data
from the Maryland State Board of Elections. The package returns tidy data
frames for easy analysis with tidyverse tools.
}
\section{Main Functions}{

\itemize{
  \item \code{\link{md_download}}: Download bulk data from the SBE API
  \item \code{\link{md_committees}}: Load committee registration data
  \item \code{\link{md_contributions}}: Load contribution and loan transactions
  \item \code{\link{md_expenditures}}: Load expenditure and IE/EC transactions
}
}

\section{Data Sources}{

All data comes from the Maryland State Board of Elections Campaign
Reporting Information System (CRIS), available at:
https://campaignfinance.maryland.gov/public/cf/downloads
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NewsAppsUMD/freestatemoney}
  \item Report bugs at \url{https://github.com/NewsAppsUMD/freestatemoney/issues}
}

}
\author{
\strong{Maintainer}: NewsAppsUMD \email{dpwillis@umd.edu}

}
\keyword{internal}
