% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_md_amount}
\alias{parse_md_amount}
\title{Parse Maryland currency format}
\usage{
parse_md_amount(x)
}
\arguments{
\item{x}{Character vector of amounts}
}
\value{
Numeric vector
}
\description{
Amounts arrive dollar-formatted, e.g. `$3,000.00`.
}
\keyword{internal}
