% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_md_date}
\alias{parse_md_date}
\title{Parse Maryland date format}
\usage{
parse_md_date(date_string)
}
\arguments{
\item{date_string}{Character string with date in MM/DD/YYYY format}
}
\value{
Date object
}
\description{
Helper function to consistently parse dates from Maryland data files
}
\keyword{internal}
